Replace the form-data email.mime* calls with manual implementation.
The email.mime* implementation always replaces newlines with CRLF
(or the chosen sep, but only CRLF is valid in MIME), even inside the
objects (if they are for example binary data), which changes the line
endings in the files sent to the NameTag service.

The new implementation keeps the line separators from the original file.
That does not violate MIME spec, because the "protocol-level" line
endings are still CRLF.